processorexample/doc-online-analyzer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.32.5
python-dotenv==1.0.0
openai==1.54.0
httpx==0.25.0
PyPDF2==3.0.1
python-docx==1.1.0
lxml==4.9.3